pyproject.toml
[tool.poetry]
name = "altvote"
version = "0.1.0"
description = ""
authors = ["David Saidov <delightxxls@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = {extras = ["with-social"], version = "6.0.0"}
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
